opt/apps/Pema_BE/requirements.txt
# FastAPI and web framework
FastAPI
uvicorn
python-multipart

# Database
sqlalchemy[asyncio]
asyncpg
alembic
aiosqlite

# Caching and async
redis[hiredis]
aiofiles

# Authentication and security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Configuration and environment
pydantic
pydantic-settings
email-validator

# HTTP client
httpx
requests

# Date and time
python-dateutil

# Email
emails

# Background tasks
celery

# Monitoring and logging
structlog

# Development and testing
pytest
pytest-asyncio
pytest-httpx

# Linting and formatting
black
isort
flake8

# Type checking
mypy

# Documentation
mkdocs
mkdocs-material

uvicorn
defusedxml
watchfiles
yarl
zipp
alembic
psycopg2-binary
